Simplified random effect sampling.
I switched the sampling and priors of the random effects back to the regular rtnorm/dtnorm forms because the '_c' forms weren't working. Turns out passing the actual mean (E[x]) and variance (Var[x]) and using these to compute appropriate parameters mu and sigma for the truncated normal is not that simple (though it is in reverse; see Wikipedia on "Truncated Normal Distribution").

Also, BUGFIX: alpha_rp1 was incorrectly computed as a product rather than a sum.